+++ src/PSRule/Commands/RuleLanguageNouns.cs
@@ -10,5 +10,6 @@
         public const string AllOf = "AllOf";
         public const string RuleDefinition = "RuleDefinition";
         public const string Recommendation = "Recommendation";
+        public const string Reason = "Reason";
     }
 }