+++ src/PSRule/PSRule.psm1
@@ -51,7 +51,7 @@ function Invoke-PSRule {
         [String]$OutputPath,
 
         [Parameter(Mandatory = $False)]
-        [ValidateSet('None', 'Yaml', 'Json', 'NUnit3', 'Csv')]
+        [ValidateSet('None', 'Yaml', 'Json', 'NUnit3', 'Csv', 'Wide')]
         [PSRule.Configuration.OutputFormat]$OutputFormat,
 
         # A list of paths to check for rule definitions
@@ -1141,6 +1141,24 @@ function TypeOf {
 
 <#
 .LINK
+https://berniewhite.github.io/PSRule/keywords/PSRule/en-US/about_PSRule_Keywords.html#reason
+#>
+function Reason {
+    [CmdletBinding()]
+    [OutputType([void])]
+    param (
+        [Parameter(Mandatory = $False, Position = 0)]
+        [String]$Text
+    )
+
+    begin {
+        # This is just a stub to improve rule authoring and discovery
+        Write-Error -Message $LocalizedHelp.KeywordOutsideEngine -Category InvalidOperation;
+    }
+}
+
+<#
+.LINK
 https://berniewhite.github.io/PSRule/keywords/PSRule/en-US/about_PSRule_Keywords.html#recommend
 #>
 function Recommend {
@@ -1148,7 +1166,7 @@ function Recommend {
     [OutputType([void])]
     param (
         [Parameter(Mandatory = $False, Position = 0)]
-        [String]$Message
+        [String]$Text
     )
 
     begin {
@@ -1535,6 +1553,7 @@ function InitEditorServices {
                 'Match'
                 'TypeOf'
                 'Within'
+                'Reason'
                 'Recommend'
             );